Bug Fix: Rare Connection Hold Error on Signup.
Holding connections for signup to reduce server load on requests caused a bug error where if a failed signup attempt was made and then the user logged into another account the connection wouldn't close on the server side.

The Fix follows the protocol for logins now, where the server forces every request to be individual.

Testing Still Required for Fix.